+++ CHANGES.md
@@ -1,5 +1,15 @@
 # @geoblocks/geoblocks changes
 
+## 0.2.15
+
+- Use ol getFontParameters to parse font.
+- Fallback to fill color if none is specified on circle style.
+
+Breaking changes:
+
+- Pass feature object to allow informed customization.
+
+
 ## 0.2.14
 
 - Fix (invert) text Y axis offset.